switch to pyproject, add build/upload action
