+++ .pre-commit-config.yaml
@@ -9,7 +9,7 @@ repos:
     -   id: mixed-line-ending
         args: ['--fix=lf']
 -   repo: https://github.com/psf/black
-    rev: 22.3.0
+    rev: 22.6.0
     hooks:
     -   id: black
 -   repo: https://gitlab.com/pycqa/flake8
@@ -32,8 +32,3 @@ repos:
       - id: isort
         name: isort (pyi)
         types: [pyi]
--   repo: https://github.com/kynan/nbstripout
-    rev: 0.3.9
-    hooks:
-    -   id: nbstripout
-        files: ".ipynb"
