--- old/deploy_docs.yml
@@ -1,35 +1,0 @@
-# Github action to build and deploy docs to GH-pages
-name: on_master_commit
-on:
-  push:
-    branches:
-      - master
-
-jobs:
-  update_master_docs:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v1
-
-      - name: Setup conda
-        uses: conda-incubator/setup-miniconda@v2
-        with:
-          miniconda-version: 'latest'
-          python-version: 3.9
-          channels: conda-forge
-
-      - name: install dascore with docbuild reqs
-        shell: bash -l {0}
-        run: |
-          conda install -c conda-forge pandoc
-          python -m pip install -e .[docs]
-
-      - name: build docs
-        shell: bash -l {0}
-        run: |
-          python scripts/make_docs.py
-
-      - name: Deploy 🚀
-        uses: JamesIves/github-pages-deploy-action@v4
-        with:
-          folder: ./docs/_build/html
