+++ setup.cfg
@@ -1,76 +1,6 @@
-[build-system]
-requires =
-    "setuptools >= 40.9.0",
-    "wheel",
-
-build-backend = "setuptools.build_meta"
-
-[metadata]
-name = dascore
-version = 0.1.0
-description =
-long_description = file: README.md, CHANGELOG.md, LICENSE
-keywords = geophysics, fiber-sensing, python, seismology
-license = GNU Lesser General Public License v3.0 or later (LGPLv3.0+)
-classifiers =
-    "Development Status :: 4 - Beta",
-    "Intended Audience :: Scientists/Engineers",
-    "License :: OSI Approved :: GNU Lesser General Public License v3 or later (LGPLv3+)",
-    "Programming Language :: Python :: 3.8",
-    "Programming Language :: Python :: 3.9",
-    "Programming Language :: Python :: 3.10",
-    "Topic :: Scientific/Engineering",
-
-python_requires = >= 3.8
-
-
-[options]
-zip_safe = False
-include_package_data = True
-packages = find:
-install_requires =
-    findiff
-    matplotlib
-    pandas
-    pooch>=1.2
-    setuptools
-    scipy
-    tables
-    xarray
-    pydantic>=1.9.0
-    rich
-
-[options.package_data]
-* = *.txt, *.rst
-hello = *.msg
-
-[options.entry_points]
-dascore.plugin.fiber_io =
-    TERRA15 = dascore.io.terra15.core:Terra15Formatter
-    PICKLE = dascore.io.pickle.core:PickleIO
-    WAV = dascore.io.wav.core:WavIO
-    DASDAE = dascore.io.dasdae.core:DASDAEIO
-
-[options.extras_require]
-docs =
-    sphinx>=1.5.1
-    numpydoc
-    sphinx_autodoc_typehints
-    typer
-    pandoc
-    myst-parser
-    myst-nb
-test =
-    pytest
-    coverage
-    pytest-cov
-    coveralls
-    twine
-    pre-commit
-    pytest-codeblocks
+; Note: We still need setup.cfg until flake8 supports pyproject.toml
 
 [flake8]
-;verbose = true
 max-complexity = 18
 ; match black line-length
 max-line-length = 88
